feat(roles, permissions): add roles and permissions to the application

+++ app/Http/Controllers/PermissionController.php
@@ -2,19 +2,36 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Account;
+use App\Traits\AccountTrait;
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Permission;
 
 class PermissionController extends Controller
 {
-    public function list()
+	use AccountTrait;
+
+    public function list(): \Illuminate\Contracts\View\View|\Illuminate\Contracts\View\Factory|\Illuminate\Foundation\Application
     {
+	    $accountId = $this->getAccountId();
+	    $account = Account::query()->find($accountId);
+
+	    if (!$account->hasPermissionTo('list_permissions')) {
+		    return view('errors.403', ['message' => 'You do not have permission to list permissions! You can ask for permission to get it.']);
+	    }
+
         $permissions = Permission::all();
-        return view('role-permission.permissions.list', compact('permissions'));
+        return view('permissions.list', compact('permissions'));
     }
     public function create(): \Illuminate\Contracts\View\View|\Illuminate\Contracts\View\Factory|\Illuminate\Foundation\Application
     {
-        return view('role-permission.permissions.create');
+	    $accountId = $this->getAccountId();
+	    $account = Account::query()->find($accountId);
+
+	    if (!$account->hasPermissionTo('create_permission')) {
+		    return view('errors.403', ['message' => 'You do not have permission to create permissions! You can ask for permission to get it.']);
+	    }
+        return view('permissions.create');
     }
 
     public function store(Request $request): \Illuminate\Http\RedirectResponse
@@ -31,8 +48,15 @@ class PermissionController extends Controller
 
     public function edit($id): \Illuminate\Contracts\View\View|\Illuminate\Contracts\View\Factory|\Illuminate\Foundation\Application
     {
+	    $accountId = $this->getAccountId();
+	    $account = Account::query()->find($accountId);
+
+	    if (!$account->hasPermissionTo('update_permission')) {
+		    return view('errors.403', ['message' => 'You do not have permission to edit permissions! You can ask for permission to get it.']);
+	    }
+
         $permission = Permission::findOrFail($id);
-        return view('role-permission.permissions.create', compact('permission'));
+        return view('permissions.create', compact('permission'));
     }
 
     public function update(Request $request, $id): \Illuminate\Http\RedirectResponse
@@ -52,6 +76,14 @@ class PermissionController extends Controller
 
     public function destroy($id): \Illuminate\Http\RedirectResponse
     {
+	    $accountId = $this->getAccountId();
+	    $account = Account::query()->find($accountId);
+
+	    if (!$account->hasPermissionTo('delete_permission')) {
+		    return redirect()->route('permissions.list')
+			    ->with('error', 'You do not have permission to delete permissions! You can ask for permission to get it.');
+	    }
+
         $permission = Permission::query()->findOrFail($id);
         $permission->delete();
         return redirect()->route('permissions.list')->with('success', 'Permission deleted successfully');
